# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neo API",
    "version": "1.0.0-rc"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/outlet/{tenant_id}/kds/queue": {
      "get": {
        "summary": "List Queue",
        "description": "Return active orders for the KDS queue view.",
        "operationId": "list_queue_api_outlet__tenant_id__kds_queue_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Queue Api Outlet  Tenant Id  Kds Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/outlet/{tenant_id}/kds/order/{order_id}/accept": {
      "post": {
        "summary": "Accept Order",
        "description": "Mark an order as accepted.",
        "operationId": "accept_order_api_outlet__tenant_id__kds_order__order_id__accept_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Accept Order Api Outlet  Tenant Id  Kds Order  Order Id  Accept Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/outlet/{tenant_id}/kds/order/{order_id}/progress": {
      "post": {
        "summary": "Progress Order",
        "description": "Move an order to ``IN_PROGRESS``.",
        "operationId": "progress_order_api_outlet__tenant_id__kds_order__order_id__progress_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Progress Order Api Outlet  Tenant Id  Kds Order  Order Id  Progress Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/outlet/{tenant_id}/kds/order/{order_id}/serve": {
      "post": {
        "summary": "Serve Order",
        "description": "Mark an order as served.",
        "operationId": "serve_order_api_outlet__tenant_id__kds_order__order_id__serve_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Serve Order Api Outlet  Tenant Id  Kds Order  Order Id  Serve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/outlet/{tenant_id}/kds/order/{order_id}/reject": {
      "post": {
        "summary": "Reject Order",
        "description": "Mark an order as rejected.",
        "operationId": "reject_order_api_outlet__tenant_id__kds_order__order_id__reject_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reject Order Api Outlet  Tenant Id  Kds Order  Order Id  Reject Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/outlet/{tenant_id}/kds/item/{order_item_id}/accept": {
      "post": {
        "summary": "Accept Item",
        "description": "Mark an order item as accepted.",
        "operationId": "accept_item_api_outlet__tenant_id__kds_item__order_item_id__accept_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "order_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Accept Item Api Outlet  Tenant Id  Kds Item  Order Item Id  Accept Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/outlet/{tenant_id}/kds/item/{order_item_id}/progress": {
      "post": {
        "summary": "Progress Item",
        "description": "Move an order item to ``IN_PROGRESS``.",
        "operationId": "progress_item_api_outlet__tenant_id__kds_item__order_item_id__progress_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "order_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Progress Item Api Outlet  Tenant Id  Kds Item  Order Item Id  Progress Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/outlet/{tenant_id}/kds/item/{order_item_id}/serve": {
      "post": {
        "summary": "Serve Item",
        "description": "Mark an order item as served.",
        "operationId": "serve_item_api_outlet__tenant_id__kds_item__order_item_id__serve_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "order_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Serve Item Api Outlet  Tenant Id  Kds Item  Order Item Id  Serve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/outlet/{tenant_id}/kds/item/{order_item_id}/reject": {
      "post": {
        "summary": "Reject Item",
        "description": "Mark an order item as rejected.",
        "operationId": "reject_item_api_outlet__tenant_id__kds_item__order_item_id__reject_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "order_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reject Item Api Outlet  Tenant Id  Kds Item  Order Item Id  Reject Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}